IA-2968 Extract hook calls related to beneficiary...
into custom hook.
